+++ packages/tsurlfilter/src/rules/cosmetic-rule.ts
@@ -5,7 +5,7 @@ import { DomainModifier } from '../modifiers/domain-modifier';
 import { hasUnquotedSubstring, indexOfAny } from '../utils/string-utils';
 import { getRelativeUrl } from '../utils/url';
 import { SimpleRegex } from './simple-regex';
-import { CosmeticRuleParser } from './cosmetic-rule-parser';
+import { CosmeticRuleParser, isUrlPatternResult } from './cosmetic-rule-parser';
 import { Request } from '../request';
 import { Pattern } from './pattern';
 import { ScriptletParser } from '../engine/cosmetic-engine/scriptlet-parser';
@@ -148,9 +148,10 @@ export class CosmeticRule implements rule.IRule {
 
     private extendedCss = false;
 
-    private readonly permittedDomains: string[] | undefined = undefined;
-
-    private readonly restrictedDomains: string[] | undefined = undefined;
+    /**
+     * $domain modifier pattern. It is only set if $domain modifier is specified for this rule.
+     */
+    private domainModifier: DomainModifier | null = null;
 
     /**
      * $path modifier pattern. It is only set if $path modifier is specified for this rule.
@@ -332,8 +333,21 @@ export class CosmeticRule implements rule.IRule {
     /**
      * Gets list of permitted domains.
      */
-    getPermittedDomains(): string[] | undefined {
-        return this.permittedDomains;
+    getPermittedDomains(): string[] | null {
+        if (this.domainModifier) {
+            return this.domainModifier.getPermittedDomains();
+        }
+        return null;
+    }
+
+    /**
+     * Gets list of restricted domains.
+     */
+    getRestrictedDomains(): string[] | null {
+        if (this.domainModifier) {
+            return this.domainModifier.getRestrictedDomains();
+        }
+        return null;
     }
 
     /**
@@ -344,14 +358,7 @@ export class CosmeticRule implements rule.IRule {
      * @return {boolean}
      */
     isGeneric(): boolean {
-        return !this.permittedDomains || this.permittedDomains.length === 0;
-    }
-
-    /**
-     * Gets list of restricted domains.
-     */
-    getRestrictedDomains(): string[] | undefined {
-        return this.restrictedDomains;
+        return !this.domainModifier?.hasPermittedDomains();
     }
 
     isExtendedCss(): boolean {
@@ -402,26 +409,18 @@ export class CosmeticRule implements rule.IRule {
         if (pattern) {
             // This means that the marker is preceded by the list of domains and modifiers
             // Now it's a good time to parse them.
-            const {
-                url,
-                path,
-                permittedDomains,
-                restrictedDomains,
-            } = CosmeticRuleParser.parseRulePattern(pattern);
+            const parsedPattern = CosmeticRuleParser.parseRulePattern(pattern);
 
-            if (url) {
-                this.urlModifier = new Pattern(url);
+            if (isUrlPatternResult(parsedPattern)) {
+                this.urlModifier = new Pattern(parsedPattern.url);
             } else {
+                const { path, domainModifier } = parsedPattern;
                 if (path || path === '') {
                     this.pathModifier = new Pattern(path);
                 }
 
-                if (permittedDomains) {
-                    this.permittedDomains = permittedDomains;
-                }
-
-                if (restrictedDomains) {
-                    this.restrictedDomains = restrictedDomains;
+                if (domainModifier) {
+                    this.domainModifier = domainModifier;
                 }
             }
         }
@@ -436,7 +435,10 @@ export class CosmeticRule implements rule.IRule {
      * @param request - request to check
      */
     match(request: Request): boolean {
-        if (!this.permittedDomains && !this.restrictedDomains && !this.pathModifier && !this.urlModifier) {
+        if (!this.domainModifier
+            && !this.pathModifier
+            && !this.urlModifier
+        ) {
             return true;
         }
 
@@ -444,20 +446,8 @@ export class CosmeticRule implements rule.IRule {
             return this.urlModifier.matchPattern(request, false);
         }
 
-        if (this.matchesRestrictedDomains(request.hostname)) {
-            /**
-             * Domain or host is restricted
-             * i.e. ~example.org##rule
-             */
-            return false;
-        }
-
-        if (this.hasPermittedDomains()) {
-            if (!DomainModifier.isDomainOrSubdomainOfAny(request.hostname, this.permittedDomains!)) {
-                /**
-                 * Domain is not among permitted
-                 * i.e. example.org##rule and we're checking example.org
-                 */
+        if (this.domainModifier) {
+            if (!this.domainModifier.matchDomain(request.hostname)) {
                 return false;
             }
         }
@@ -593,37 +583,6 @@ export class CosmeticRule implements rule.IRule {
         if (ruleContent.indexOf('\\', ruleContent.lastIndexOf('{')) > -1) {
             throw new SyntaxError('CSS injection rule with \'\\\' was omitted');
         }
-    }
-
-    /**
-     * Checks if the rule has permitted domains
-     */
-    private hasPermittedDomains(): boolean {
-        return this.permittedDomains != null && this.permittedDomains.length > 0;
-    }
-
-    /**
-     * Checks if the rule has restricted domains
-     */
-    private hasRestrictedDomains(): boolean {
-        return this.restrictedDomains != null && this.restrictedDomains.length > 0;
-    }
-
-    /**
-     * Checks if the hostname matches permitted domains
-     * @param hostname
-     */
-    public matchesPermittedDomains(hostname: string): boolean {
-        return this.hasPermittedDomains() && DomainModifier.isDomainOrSubdomainOfAny(hostname, this.permittedDomains!);
-    }
-
-    /**
-     * Checks if the hostname matches the restricted domains.
-     * @param hostname
-     */
-    public matchesRestrictedDomains(hostname: string): boolean {
-        return this.hasRestrictedDomains()
-            && DomainModifier.isDomainOrSubdomainOfAny(hostname, this.restrictedDomains!);
     }
 
     /**