+++ packages/tsurlfilter/test/modifiers/domain-modifier.test.ts
@@ -46,6 +46,13 @@ describe('Domain modifier', () => {
                     restricted: ['example.com'],
                 },
             },
+            {
+                actual: 'example.org,~example.com,example.*,/io/|~/net/',
+                expected: {
+                    permitted: ['example.org', 'example.*', '/io/|~/net/'],
+                    restricted: ['example.com'],
+                },
+            },
         ];
         test.each(domainsListCases)('%s', ({ actual, expected }) => {
             const domainModifier = new DomainModifier(actual, COMMA_SEPARATOR);
@@ -123,6 +130,8 @@ describe('Domain modifier', () => {
     describe('constructor and invalid domains', () => {
         const COMMA_SEPARATOR = ',';
         const EMPTY_DOMAIN_ERROR = 'Empty domain specified in';
+        const STARTS_WITH_SEPARATOR_ERROR = `Modifier $domain cannot start with "${COMMA_SEPARATOR}"`;
+        const HAS_INVALID_WILDCARD = 'Wildcards are only supported for top-level domains:';
         const invalidCases = [
             {
                 actual: '',
@@ -141,15 +150,7 @@ describe('Domain modifier', () => {
                 error: EMPTY_DOMAIN_ERROR,
             },
             {
-                actual: ',',
-                error: EMPTY_DOMAIN_ERROR,
-            },
-            {
                 actual: 'example.com,',
-                error: EMPTY_DOMAIN_ERROR,
-            },
-            {
-                actual: ',example.com',
                 error: EMPTY_DOMAIN_ERROR,
             },
             {
@@ -164,6 +165,18 @@ describe('Domain modifier', () => {
                 actual: 'example.com,  ,example.org',
                 error: EMPTY_DOMAIN_ERROR,
             },
+            {
+                actual: ',example.com',
+                error: STARTS_WITH_SEPARATOR_ERROR,
+            },
+            {
+                actual: ',',
+                error: STARTS_WITH_SEPARATOR_ERROR,
+            },
+            {
+                actual: 'example.com,*.org',
+                error: HAS_INVALID_WILDCARD,
+            },
         ];
         test.each(invalidCases)('%s', ({ actual, error }) => {
             expect(() => {
@@ -172,24 +185,39 @@ describe('Domain modifier', () => {
         });
     });
 
-    it('works in common cases', () => {
-        expect(DomainModifier.isDomainOrSubdomainOfAny('example.org', ['example.org'])).toBeTruthy();
+    describe('DomainModifier.isDomainOrSubdomainOfAny', () => {
+        const { isDomainOrSubdomainOfAny } = DomainModifier;
+        it('works in common cases', () => {
+            expect(isDomainOrSubdomainOfAny('example.org', ['example.org'])).toBeTruthy();
 
-        expect(DomainModifier.isDomainOrSubdomainOfAny('example.com', ['example.org'])).toBeFalsy();
-        expect(DomainModifier.isDomainOrSubdomainOfAny('', ['example.org'])).toBeFalsy();
-        expect(DomainModifier.isDomainOrSubdomainOfAny('example.org', [])).toBeFalsy();
+            expect(isDomainOrSubdomainOfAny('example.com', ['example.org'])).toBeFalsy();
+            expect(isDomainOrSubdomainOfAny('', ['example.org'])).toBeFalsy();
+            expect(isDomainOrSubdomainOfAny('example.org', [])).toBeFalsy();
+        });
+
+        it('works in wildcard cases', () => {
+            expect(isDomainOrSubdomainOfAny('example.org', ['example.*', 'test.com'])).toBeTruthy();
+            expect(isDomainOrSubdomainOfAny('sub.example.org', ['example.*', 'test.com'])).toBeTruthy();
+            expect(isDomainOrSubdomainOfAny(
+                'example.org',
+                ['one.*', 'example.*', 'test.com'],
+            )).toBeTruthy();
+            expect(isDomainOrSubdomainOfAny('www.chrono24.ch', ['chrono24.*'])).toBeTruthy();
+
+            expect(isDomainOrSubdomainOfAny('example.com', ['test.*'])).toBeFalsy();
+            expect(isDomainOrSubdomainOfAny('subexample.org', ['example.*'])).toBeFalsy();
+            expect(isDomainOrSubdomainOfAny('example.eu.uk', ['example.*'])).toBeFalsy();
+            expect(isDomainOrSubdomainOfAny('example.org', ['sub.example.*', 'test.com'])).toBeFalsy();
+            expect(isDomainOrSubdomainOfAny('', ['example.*', 'test.com'])).toBeFalsy();
+        });
     });
 
-    it('works in wildcard cases', () => {
-        expect(DomainModifier.isDomainOrSubdomainOfAny('example.org', ['example.*', 'test.com'])).toBeTruthy();
-        expect(DomainModifier.isDomainOrSubdomainOfAny('sub.example.org', ['example.*', 'test.com'])).toBeTruthy();
-        expect(DomainModifier.isDomainOrSubdomainOfAny('example.org', ['one.*', 'example.*', 'test.com'])).toBeTruthy();
-        expect(DomainModifier.isDomainOrSubdomainOfAny('www.chrono24.ch', ['chrono24.*'])).toBeTruthy();
-
-        expect(DomainModifier.isDomainOrSubdomainOfAny('example.com', ['test.*'])).toBeFalsy();
-        expect(DomainModifier.isDomainOrSubdomainOfAny('subexample.org', ['example.*'])).toBeFalsy();
-        expect(DomainModifier.isDomainOrSubdomainOfAny('example.eu.uk', ['example.*'])).toBeFalsy();
-        expect(DomainModifier.isDomainOrSubdomainOfAny('example.org', ['sub.example.*', 'test.com'])).toBeFalsy();
-        expect(DomainModifier.isDomainOrSubdomainOfAny('', ['example.*', 'test.com'])).toBeFalsy();
+    describe('DomainModifier.isNonPlainDomain', () => {
+        const { isWildcardOrRegexDomain } = DomainModifier;
+        it('distinguishes plain domains from patterns', () => {
+            expect(isWildcardOrRegexDomain('example.co.uk')).toBeFalsy();
+            expect(isWildcardOrRegexDomain('example.*')).toBeTruthy();
+            expect(isWildcardOrRegexDomain(String.raw`/another\.(org|com)/`)).toBeTruthy();
+        });
     });
 });