+++ packages/tsurlfilter/test/rules/cosmetic-rule-parser.test.ts
@@ -1,4 +1,5 @@
 import { CosmeticRuleParser } from '../../src/rules/cosmetic-rule-parser';
+import { DomainModifier, COMMA_SEPARATOR, PIPE_SEPARATOR } from '../../src/modifiers/domain-modifier';
 
 describe('CosmeticRuleParser', () => {
     it('parse rule text by marker', () => {
@@ -90,34 +91,37 @@ describe('CosmeticRuleParser', () => {
     });
 
     it('parse and extracts the permitted/restricted domains and the unescaped path modifier value', () => {
-        expect(CosmeticRuleParser.parseRulePattern('example.org')).toEqual({
-            permittedDomains: ['example.org'],
+        let rulePattern: string;
+
+        rulePattern = 'example.org';
+        expect(CosmeticRuleParser.parseRulePattern(rulePattern)).toEqual({
+            domainModifier: new DomainModifier(rulePattern, COMMA_SEPARATOR),
         });
 
-        expect(CosmeticRuleParser.parseRulePattern('example.org,another.com')).toEqual({
-            permittedDomains: ['example.org', 'another.com'],
+        rulePattern = 'example.org,another.com';
+        expect(CosmeticRuleParser.parseRulePattern(rulePattern)).toEqual({
+            domainModifier: new DomainModifier(rulePattern, COMMA_SEPARATOR),
         });
 
         expect(CosmeticRuleParser.parseRulePattern('*')).toEqual({});
 
-        expect(CosmeticRuleParser.parseRulePattern('example.org,~another.com')).toEqual({
-            permittedDomains: ['example.org'],
-            restrictedDomains: ['another.com'],
+        rulePattern = 'example.org,~another.com';
+        expect(CosmeticRuleParser.parseRulePattern(rulePattern)).toEqual({
+            domainModifier: new DomainModifier(rulePattern, COMMA_SEPARATOR),
         });
 
-        expect(CosmeticRuleParser.parseRulePattern('[$url=example.com/category/5]')).toEqual({
-            url: 'example.com/category/5',
+        rulePattern = 'example.com/category/5';
+        expect(CosmeticRuleParser.parseRulePattern(`[$url=${rulePattern}]`)).toEqual({
+            url: rulePattern,
         });
 
         expect(CosmeticRuleParser.parseRulePattern('[$path=/page]example.org,~another.com')).toEqual({
-            permittedDomains: ['example.org'],
-            restrictedDomains: ['another.com'],
+            domainModifier: new DomainModifier('example.org,~another.com', COMMA_SEPARATOR),
             path: '/page',
         });
 
         expect(CosmeticRuleParser.parseRulePattern('[$path=/page,domain=example.org|~another.com]')).toEqual({
-            permittedDomains: ['example.org'],
-            restrictedDomains: ['another.com'],
+            domainModifier: new DomainModifier('example.org|~another.com', PIPE_SEPARATOR),
             path: '/page',
         });